refactor : api cud/r Read파일 Query파일로 분리
