+++ src/main/java/com/rualone/app/domain/follow/application/FollowService.java
@@ -11,5 +11,4 @@ public interface FollowService {
 
     void create(FollowCreateRequest followCreateRequest);
 
-    List<Follow> read(Long followee);
 }
