+++ src/main/java/com/rualone/app/domain/follow/dto/response/FollowReadResponse.java
@@ -1,0 +1,12 @@
+package com.rualone.app.domain.follow.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class FollowReadResponse {
+    private Long id;
+    private Long follower;
+    private Long followee;
+}
